Use php-i18n via Composer, override language detection as WFi18n
The latest version of php-i18n puts `$_COOKIE['lang']` at the bottom of
the detection logic, meaning that the value of that cookie will not be
used if HTTP_ACCEPT_LANGUAGE is set - exactly what we don't want - so
this subclasses \i18n and overrides the language selection function
(getUserLangs).

This makes things slightly more maintainable - we can upgrade the i18n
library in future, but still keep the language selection as we want it.