Yeet `require_once` from `loader.php`, use Composer PSR-0 autoload instead
This brings a bunch of benefits - most obviously, there's no manual
maintenance of the list of requires in `loader.php` anymore. It also
might make load times _slightly_ faster, because classes will only get
loaded if they are actually used by the page that's currently being
requested (rather than just "load all the things") - but I have not
benchmarked this, and honestly, there's no need to.

Because PSR-0 is basically just a way for Composer to create a file
saying "here's the file that each of these classes is located in,"
(rather than PSR-4's fancy "I know the name of this class and I can
derive a filename from that"), Composer's PSR-0 cache must be updated
if you add or rename something in `src/` - you can do this by running
`composer dumpautoload`.

The autoload files are automatically generated when running either of
`composer install` / `composer update`, so as long as you're doing that
when you deploy the site to production, the process for deployment does
not change with this commit.