WFHtmlSanitizer: Add working HtmlSanitizer extension
This implements a configurable class attribute override (with an option
to preserve existing classes and just prepend our configured ones) for
<img> elements, which builds on top of the default "image" extension
(\HtmlSanitizer\Extension\Image) for HtmlSanitizer.

The extension is available as \WFHtmlSanitizer\WFExtension and can be
registered using the \HtmlSanitizer\SanitizerBuilder interface, see the
documentation at the top of WFExtension.php for details.