WFText: Remove `makeTextRenderableForEdit` call from `makeTextStripped`
This changes `makeTextStripped` to run Parsedown on the input directly,
rather than going through the "normal" post HTML rendering functions,
which makes it a bit quicker - shorttag replacement (mentions, images,
and site emoji) and the "big" HtmlSanitizer (which goes through and
modifies a bunch of things) don't run.

Nothing that is done in the call to `makeTextRenderableForEdit` aside
from the Parsedown rendering is needed in this case, because literally
all of it gets removed by the HtmlSanitizer in `makeTextStripped`.

This change also replaces all line breaks in the HtmlSanitizer output
with a single space character, improving rendering in OpenGraph tags.