WFText: Rewrite and rename `mentionReplace` to `shorttagMentionRender`
I will admit, this rewrite of a previously simple function looks a bit
daunting, but it's far safer in terms of untrusted input as well as
constructing the links that are sent as part of the sanitized HTML.

It's pretty well commented, so you should be able to make sense of it -
the only weird thing is the \DOMDocument stuff. That creates the HTML
<a> element programmatically in a safe way, rather than doing string
interpolation to create these links.

This commit also adds the `unidentified-blog` URL to the list of
disallowed blog names in `WFUtils::blogNameCheck`, as that URL is now
used for rendering MENTION shorttags where retrieving the mentioned
blog by it's ID was not successful.

A note for post editing: To implement handling of mentions properly,
the edit page MUST call `makeTextRenderableForEdit` before passing the
post content to the browser to show in the editor (already done), but
MUST ALSO convert the <a data-url-mentions=""> back into the MENTION
shorttag before saving the edited post to the database. This will be
implemented in a follow-up commit.