WFText: Rewrite `makeTextSafe`
This changes the purpose of `makeTextSafe` - it has become a "strip all
the things" implementation, suitable for use for page/post titles, image
IDs/captions, that sorta thing - which is largely how the old function
was used through the codebase.

The exception to this is post/page content sanitization where the data
comes from untrusted sources (such as the text editor on the site) -
for those use cases, a new `makeTextPostContentSafe` function will be
implemented. That new (as yet unimplemented) function is mentioned in
the documentation comment of `makeTextSafe`, so hopefully people read
that.