WFHtmlSanitizer: Add HTML <img> to {{IMAGE:…}} shortcode conversion
This is the first of a series of commits that will convert HTML tags
back into shortcodes for the purposes of the content being submitted
through a post or page edit form, the input sanitization of which will
be handled by the upcoming `WFText::makeTextPostContentSafe` function.

This adds the `\WFHtmlSanitizer\ShortcodeNode` node type designed for
use within a NodeVisitor, to replace the node being processed by that
NodeVisitor with a shortcode.

This also adds the `waterfall-shortcodes` section to the HtmlSanitizer
configuration. This can contain an `enabled` key, which is an array of
the types of shortcode conversion to enable for the run. The conversion
added in this commit can be enabled with `'image'`, like so:

    $sanitizer = $sanitizerBuilder->build([
        'extensions' => ['waterfall'],
        'waterfall-shortcodes' => ['enabled' => ['image']],
    ]);

A single test has been implemented for this functionality, and more
in-depth unit testing for the shortcode features of `\WFHtmlSanitizer`
is planned.