WFText: Rewrite &  rename `doReadMoreCheck` to `shorttagReadMoreRender`
The new `shorttagReadMoreRender` function uses the principle that the
`shorttagMentionRender` function uses to generate links, but applied
to the Read More button _and_ the container that holds the hidden post
content.

The `shorttagReadMoreRender` function splits the post segment by the
read more shortcode (`{{READMORE}}`), runs an `HtmlSanitizer` pass on
both halves of the split, loads those two halves as individual UTF-8
encoded `DOMDocument` instances (named `$contentPreDoc` for the content
before the marker, and `$contentPostDoc` for the content after),
programmatically creates the "Read More" button (which is now able to
be translated) and the container that holds the `$contentPostDoc`
data, and then returns all of that serialized back to HTML.

Unfortunately, as with the `shorttagMentionRender` function, this
change turns the function that was previous 9 lines long into somewhat
of a monstrosity - but I'm confident that doing the creation of the
HTML elements by this method is going to be far better in the long run.