WFText: Rewrite & rename `imageReplace` to `shorttagImageRender`
This, again, takes the virtual DOM approach to creating HTML elements
for rendering in the browser. This is especially important for this
portion of the code, because the image shortcode rendering embeds user-
provided text directly into the element attributes, which was done by
string interpolation in the previous `imageReplace` function.

The <img> element generated by the new function will always use a
downsized version of the source image (defined by the `$deviceType`
detection) as the `src="…"` attribute, however the image that is set as
the `href="…"` of the surrounding <a> element is always the full-size
image. This was done to simplify the logic ever so slightly, in a way
that 99% of users will never notice, because they'll tap on the image
(which will load the full-res version in the lightbox).